# specs/5-better-auth-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Better Auth Integration API
  description: API specification for Better Auth (JWT) integration with Todo app
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with Better Auth
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "securePassword123"
              required:
                - email
                - password
      responses:
        '200':
          description: User registered successfully, JWT token set in httpOnly cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "better-auth-session=eyJhb...; HttpOnly; Secure; SameSite=Strict"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      summary: Login existing user
      description: Authenticates user and sets JWT token in httpOnly cookie
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "securePassword123"
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful, JWT token set in httpOnly cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "better-auth-session=eyJhb...; HttpOnly; Secure; SameSite=Strict"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      summary: Logout current user
      description: Clears the JWT token and ends the session
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful, JWT token cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth, passed in Authorization header or retrieved from httpOnly cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
      required:
        - id
        - email
        - created_at
        - updated_at
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Bad Request"
              message:
                type: string
                example: "Invalid input parameters"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication token is missing or invalid"
...
    Conflict:
      description: Resource conflict (e.g., duplicate email)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Conflict"
              message:
                type: string
                example: "Resource already exists"